requirements.txt
email-validator==2.1.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
